% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggRate.R
\name{aggRate}
\alias{aggRate}
\alias{aggRate.data.frame}
\alias{aggRate.formula}
\alias{aggRate.clm}
\alias{aggRate.vlm}
\title{Create Aggregated 'Flat' Format of Rating Data}
\usage{
aggRate(object, ...)

\method{aggRate}{data.frame}(object, pattern, rating, ...)

\method{aggRate}{formula}(formula, data, ...)

\method{aggRate}{clm}(object, ...)

\method{aggRate}{vlm}(object, ...)
}
\arguments{
\item{object}{..}

\item{...}{..}

\item{pattern, rating}{..}

\item{formula}{..}

\item{data}{..}
}
\description{
Function \link{aggRate} creates aggregate counts rating data \emph{flat} format,
for \link[VGAM]{vglm} as well as human eyes.
Each row represents the aggregated rating records for a single \emph{product},
and each column denotes the levels of rating (e.g. \link[VGAM]{wine}).
}
\examples{
VGAM::wine |>
 aggRate.data.frame(pattern = c(bitter = '^bitter'))

ordinal::wine |> 
 aggRate.formula(formula = rating ~ temp + contact)
 
m = MASS::housing |>
 reshape2::dcast(formula = Infl + Type + Cont ~ Sat, value.var = 'Freq') |>
 aggRate.data.frame(rating = list(Sat = c('Low', 'Medium', 'High')))
# printing method has a bug.

library(ordinal)
clm(rating ~ temp * contact, data = wine) |>
 aggRate.clm()
# aggRate(housing_clm) # I plan to deprecate

library(VGAM)
pneumo2 = VGAM::pneumo |>
 aggRate.data.frame(rating = list(y = c('normal', 'mild', 'severe')))
(m = vglm(y ~ log(exposure.time), family = multinomial, data = pneumo2, model = TRUE))
m |> aggRate.vlm()
m |> xlevels.vlm()

}
