% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autolayer.aggRate}
\alias{autolayer.aggRate}
\title{autolayer.aggRate}
\usage{
\method{autolayer}{aggRate}(object, type = c("plain", "cumulative", "acat", "cratio"), ...)
}
\arguments{
\item{object}{..}

\item{type}{..}

\item{...}{..}
}
\description{
..
}
