% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm.R
\name{clm_clmm}
\alias{clm_clmm}
\title{\link[ordinal]{clm} and \link[ordinal]{clmm} Objects}
\description{
\link[ordinal]{clm} and \link[ordinal]{clmm} Objects
}
\examples{
library(ordinal)
?ordinal::clm
?ordinal::clmm
m1 = clm(rating ~ temp + contact, data = wine)
m2 = clmm(rating ~ temp + contact + (1|judge), data = wine)
m3 = clm(Sat ~ Infl + Type + Cont, weights = Freq, data = MASS::housing)
library(ecip); list(
 '`clm`' = list(m1, m3),
 '`clmm`' = m2
) |> fastmd::render2html()
}
