% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlm.R
\name{family.vlm}
\alias{family.vlm}
\title{Get Family from \link[VGAM]{vlm} Object}
\usage{
\method{family}{vlm}(object, ...)
}
\arguments{
\item{object}{\link[VGAM]{vlm} object}

\item{...}{additional parameters, currently not in use}
}
\value{
The \code{S3} method \code{\link[=family.vlm]{family.vlm()}} returns a \code{'vglmff'} object.
}
\description{
Get Family from \link[VGAM]{vlm} Object
}
\examples{
library(VGAM)
pneumo = transform(pneumo, let = log(exposure.time))
(m1 = vglm(cbind(normal, mild, severe) ~ let, family = multinomial, data = pneumo))
(m2 = vglm(cbind(normal, mild, severe) ~ let, family = propodds, data = pneumo))
(m3 = vglm(cbind(normal, mild, severe) ~ let, family = acat, data = pneumo))
m1 |> desc_.vlm()
m2 |> desc_.vlm()
m3 |> desc_.vlm()

library(ecip); list(
 '`vglm`' = m2
) |> fastmd::render2html()
}
