% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlm.R
\name{md_.vlm}
\alias{md_.vlm}
\title{R Markdown Lines of \link[VGAM]{vlm} object}
\usage{
\method{md_}{vlm}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines of \link[VGAM]{vlm} object
}
\examples{
library(VGAM); library(ecip)
pneumo = transform(pneumo, let = log(exposure.time))
list(
 '`vglm`' = vglm(cbind(normal, mild, severe) ~ let, propodds, data = pneumo)
) |> fastmd::render_(file = 'vglm')

}
\keyword{internal}
