% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm.R
\name{md_ordinal}
\alias{md_ordinal}
\alias{md_.clm}
\alias{md_.clmm}
\title{R Markdown Lines for \link[ordinal]{clm} and \link[ordinal]{clmm} Objects}
\usage{
\method{md_}{clm}(x, xnm, ...)

\method{md_}{clmm}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[ordinal]{clm} and \link[ordinal]{clmm} Objects
}
\examples{
library(ordinal); library(ecip)
list(
 '`clm`' = clm(rating ~ temp + contact, data = ordinal::wine), 
 '`clmm`' = clmm(rating ~ temp + contact + (1|judge), data = ordinal::wine)
) |> rmd.tzh::render_(file = 'clm_clmm')

}
