% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_VGAM.R
\name{S3_VGAM}
\alias{S3_VGAM}
\alias{summary.vglm}
\alias{summary.vlm}
\alias{model.frame.vlm}
\title{Missing S3 Methods in \pkg{VGAM} Package}
\usage{
\method{summary}{vglm}(object, ...)

\method{summary}{vlm}(object, ...)

\method{model.frame}{vlm}(formula, ...)
}
\arguments{
\item{object, formula}{a \link[VGAM]{vlm} or \link[VGAM]{vglm} object}

\item{...}{S3 method dispatch holder}
}
\description{
Missing S3 Methods in \pkg{VGAM} Package
}
\details{
tzh has written to Dr. Yee about some of these ..
}
