% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlm.R
\name{S3_vglmff}
\alias{S3_vglmff}
\alias{getCanonicalLink.vglmff}
\alias{getLink.vglmff}
\title{S3 Methods for \code{vglmff} Object}
\usage{
\method{getCanonicalLink}{vglmff}(x)

getLink.vglmff(x)
}
\arguments{
\item{x}{\code{vglmff} family object}
}
\description{
S3 Methods for \code{vglmff} Object
}
