% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlm.R
\name{S3_vlm}
\alias{S3_vlm}
\alias{.pval.vglm}
\alias{.pval.summary.vglm}
\title{S3 methods of \link[VGAM]{vlm} and \link[VGAM]{vglm} Objects}
\usage{
.pval.vglm(x)

.pval.summary.vglm(x)
}
\arguments{
\item{x}{\link[VGAM]{vlm} or \link[VGAM]{vglm} object}
}
\description{
S3 methods of \link[VGAM]{vlm} and \link[VGAM]{vglm} Objects
}
