% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglm_fams.R
\name{vglm_fams}
\alias{vglm_fams}
\alias{vglm_fams.clm}
\alias{vglm_fams.aggRate}
\title{Handy extensions of \link[VGAM]{vglm}}
\usage{
vglm_fams(object, ...)

\method{vglm_fams}{clm}(object, ...)

\method{vglm_fams}{aggRate}(object, ...)
}
\arguments{
\item{object}{see \strong{Usage}}

\item{...}{see \link[VGAM]{vglm}}
}
\description{
To try multiple family choices and select the one with the least \code{ResSS}.
}
\examples{
m = VGAM::wine |>
 aggRate.data.frame(pattern = c(bitter = '^bitter')) |>
 vglm_fams()
autoplot(m, type = 'plain')

library(ordinal)
clm(rating ~ temp + contact, data = ordinal::wine) |>
 vglm_fams()
}
