% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlm.R
\name{vlm}
\alias{vlm}
\title{\link[VGAM]{vlm} Object}
\description{
\link[VGAM]{vlm} Object
}
\examples{
library(VGAM)
pneumo = transform(pneumo, let = log(exposure.time))
(m1 = vglm(cbind(normal, mild, severe) ~ let, family = multinomial, data = pneumo))
(m2 = vglm(cbind(normal, mild, severe) ~ let, family = propodds, data = pneumo))
(m3 = vglm(cbind(normal, mild, severe) ~ let, family = acat, data = pneumo))
library(ecip); list(
 '`multinomial`' = m1,
 '`propodds`' = m2,
 '`acat`' = m3
) |> fastmd::render2html()
}
