% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlm.R
\name{family.vlm}
\alias{family.vlm}
\title{Get Family from \link[VGAM]{vlm} Object}
\usage{
\method{family}{vlm}(object, ...)
}
\arguments{
\item{object}{\link[VGAM]{vlm} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{family.vlm} returns a \code{'vglmff'} object.
}
\description{
Get Family from \link[VGAM]{vlm} Object
}
