% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_VGAM.R
\name{S3_VGAM}
\alias{S3_VGAM}
\alias{summary.vlm}
\title{Missing S3 Methods in \pkg{VGAM} Package}
\usage{
\method{summary}{vlm}(object, ...)
}
\arguments{
\item{object}{a \link[VGAM]{vlm} object}

\item{...}{S3 method dispatch holder}
}
\description{
Missing S3 Methods in \pkg{VGAM} Package
}
\details{
tzh forgot to write to Dr. Yee about \code{\link[=summary.vlm]{summary.vlm()}}..
}
